Refactor build and testing
